+++ src/index.js
@@ -283,6 +283,46 @@ async function getSectionNamesForWork (work, language) {
 /**
  * @param {string} work
  * @param {"fa"|"en"} [language] If none is provided, will check all languages
+ * @returns {Promise<string[]>}
+ */
+async function getSectionInfoForWork (work, language) {
+  const [works, sections] = await Promise.all([
+    getWorks(language),
+    getSections(language)
+  ]);
+  // Due to a mismatch between the title found on the Table of Contents,
+  //   and the one found as a heading on the page (e.g.,
+  //   'سراپردۀ يگانگی' and 'سراپردهٔ یگانگی', we need to match another way
+  //   between section and work. But since the Table of Contents
+  //   apparently doesn't store the ID, we need to check the works,
+  //   directly, cross-referenced by parentUrl/url
+  const found = works.find(({title}) => {
+    return work === title;
+  });
+  /* c8 ignore next 3 */
+  if (!found) {
+    return [];
+  }
+  const subSectionMatches = sections.subSections.filter(({parentUrl}) => {
+    const mainSection = sections.mainSections.find(({
+      parentUrl: mainSectionParentUrl
+    }) => {
+      return mainSectionParentUrl === found.url;
+    });
+    return mainSection && mainSection.parentUrl === parentUrl;
+  });
+  return (
+    subSectionMatches.length
+      ? subSectionMatches
+      : sections.mainSections.filter(({parentUrl}) => {
+        return parentUrl === found.url;
+      })
+  );
+}
+
+/**
+ * @param {string} work
+ * @param {"fa"|"en"} [language] If none is provided, will check all languages
  * @returns {Promise<string>}
  */
 async function getUrlForWork (work, language) {
@@ -373,6 +413,7 @@ export {
   getUrlForId,
   getWorkNames,
   getSectionNamesForWork,
+  getSectionInfoForWork,
   getUrlForWork,
   getSubsectionUrlForWork,
   getUrlForWorkAndSection